# server/src/main/resources/static/openapi.yml
openapi: 3.0.3
info:
  title: Voz Cidadã API
  description: API para o sistema Voz Cidadã
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Servidor de desenvolvimento local
...
paths:
...
  /api/funcionario:
    get:
      tags:
        - Funcionário
      summary: Listar todos os funcionários
      description: Retorna uma lista paginada de funcionários (apenas para owners)
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/pageSizeParam'
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/sortParam'
      responses:
        '200':
          description: Lista de funcionários recuperada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/FuncionarioDTO'
                  page:
                    $ref: '#/components/schemas/PageInfo'
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão para acessar o recurso
    post:
      tags:
        - Funcionário
      summary: Criar perfil de administrador
      description: Cria um novo perfil de administrador (apenas para owners)
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FuncionarioDTO'
      responses:
        '201':
          description: Perfil de administrador criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FuncionarioDTO'
        '400':
          description: Dados de entrada inválidos
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão para criar administrador
  /api/funcionario/{id}:
    get:
      tags:
        - Funcionário
      summary: Buscar funcionário por ID
      description: Retorna os detalhes de um funcionário específico
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Funcionário encontrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FuncionarioDTO'
        '401':
          description: Não autenticado
        '404':
          description: Funcionário não encontrado
    delete:
      tags:
        - Funcionário
      summary: Excluir funcionário
      description: Exclui um funcionário específico
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Funcionário excluído com sucesso
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão para excluir este funcionário
        '404':
          description: Funcionário não encontrado
  /api/funcionario/auth/{authUserId}:
    get:
      tags:
        - Funcionário
      summary: Buscar funcionário por ID de usuário autenticado
      description: Retorna os detalhes do funcionário associado ao ID de usuário autenticado
      security:
        - bearerAuth: [ ]
      parameters:
        - name: authUserId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Funcionário encontrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FuncionarioDTO'
        '401':
          description: Não autenticado
        '404':
          description: Funcionário não encontrado

  # Chamado endpoints
  /api/chamado:
    get:
      tags:
        - Chamado
      summary: Listar todos os chamados
      description: Retorna uma lista paginada de todos os chamados
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/pageSizeParam'
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/sortParam'
      responses:
        '200':
          description: Lista de chamados recuperada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChamadoDTO'
                  page:
                    $ref: '#/components/schemas/PageInfo'
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão para acessar o recurso
    post:
      tags:
        - Chamado
      summary: Criar um novo chamado
      description: Cria um novo chamado com os dados fornecidos
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChamadoDTO'
      responses:
        '201':
          description: Chamado criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChamadoDTO'
        '400':
          description: Dados de entrada inválidos
        '401':
          description: Não autenticado
  /api/chamado/{id}:
    get:
      tags:
        - Chamado
      summary: Buscar chamado por ID
      description: Retorna os detalhes de um chamado específico
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Chamado encontrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChamadoDTO'
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão para acessar o recurso
        '404':
          description: Chamado não encontrado
    put:
      tags:
        - Chamado
      summary: Atualizar um chamado existente
      description: Atualiza os dados de um chamado existente (apenas para administradores)
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChamadoDTO'
      responses:
        '200':
          description: Chamado atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChamadoDTO'
        '400':
          description: Dados de entrada inválidos
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão para atualizar este chamado
        '404':
          description: Chamado não encontrado
    delete:
      tags:
        - Chamado
      summary: Excluir um chamado
      description: Exclui um chamado específico
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Chamado excluído com sucesso
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão para excluir este chamado
        '404':
          description: Chamado não encontrado
  /api/chamado/user/{userId}:
    get:
      tags:
        - Chamado
      summary: Listar chamados por ID de usuário
      description: Retorna uma lista paginada de chamados criados por um usuário específico
      security:
        - bearerAuth: [ ]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/pageSizeParam'
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/sortParam'
      responses:
        '200':
          description: Lista de chamados do usuário recuperada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChamadoDTO'
                  page:
                    $ref: '#/components/schemas/PageInfo'
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão para acessar chamados de outro usuário
        '404':
          description: Usuário não encontrado
  /api/chamado/secretaria/{secretaria}:
    get:
      tags:
        - Chamado
      summary: Listar chamados por secretaria
      description: Retorna uma lista paginada de chamados designados a uma secretaria específica
      security:
        - bearerAuth: [ ]
      parameters:
        - name: secretaria
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/pageSizeParam'
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/sortParam'
      responses:
        '200':
          description: Lista de chamados da secretaria recuperada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChamadoDTO'
                  page:
                    $ref: '#/components/schemas/PageInfo'
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão para acessar chamados desta secretaria

  # Upload endpoints
...
  /api/avaliacao:
    post:
      tags:
        - Avaliação
      summary: Criar uma nova avaliação
      description: Cria uma nova avaliação com os dados fornecidos
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvaliacaoDTO'
      responses:
        '201':
          description: Avaliação criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvaliacaoDTO'
          headers:
            Location:
              description: URL da avaliação criada
              schema:
                type: string
        '400':
          description: Dados de entrada inválidos
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão para criar avaliação
    get:
      tags:
        - Avaliação
      summary: Listar todas as avaliações
      description: Retorna uma lista paginada de todas as avaliações
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/pageSizeParam'
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/sortParam'
      responses:
        '200':
          description: Lista de avaliações recuperada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/AvaliacaoDTO'
                  page:
                    $ref: '#/components/schemas/PageInfo'
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão para acessar o recurso
  /api/avaliacao/{id}:
    get:
      tags:
        - Avaliação
      summary: Buscar avaliação por ID
      description: Retorna os detalhes de uma avaliação específica
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Avaliação encontrada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvaliacaoDTO'
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão para acessar o recurso
        '404':
          description: Avaliação não encontrada
    put:
      tags:
        - Avaliação
      summary: Atualizar uma avaliação existente
      description: Atualiza os dados de uma avaliação existente
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvaliacaoDTO'
      responses:
        '200':
          description: Avaliação atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvaliacaoDTO'
        '400':
          description: Dados de entrada inválidos
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão para atualizar esta avaliação
        '404':
          description: Avaliação não encontrada
    delete:
      tags:
        - Avaliação
      summary: Excluir uma avaliação
      description: Exclui uma avaliação específica
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Avaliação excluída com sucesso
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão para excluir esta avaliação
        '404':
          description: Avaliação não encontrada
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    pageSizeParam:
      name: size
      in: query
      description: Número de itens por página
      schema:
        type: integer
        default: 10
    pageNumberParam:
      name: page
      in: query
      description: Número da página (começando em 0)
      schema:
        type: integer
        default: 0
    sortParam:
      name: sort
      in: query
      description: Ordenação (ex. campo,asc ou campo,desc)
      schema:
        type: string
  schemas:
...
    AvaliacaoDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        chamadoId:
          type: integer
          format: int64
          description: ID do chamado associado à avaliação
        usuarioId:
          type: integer
          format: int64
          description: ID do usuário que criou a avaliação
        estrelas:
          type: integer
          description: Nota da avaliação em estrelas (1-5)
          minimum: 1
          maximum: 5
        comentario:
          type: string
          description: Comentário sobre a avaliação
          nullable: true
        dataAvaliacao:
          type: string
          format: date-time
          description: Data e hora da avaliação no formato 'yyyy-MM-dd HH:mm:ss'
          example: "2023-12-31 23:59:59"
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'

    # Chamado DTO
    ChamadoDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        usuarioId:
          type: integer
          format: int64
          description: ID do usuário que criou o chamado
        usuarioNome:
          type: string
          description: Nome do usuário que criou o chamado
        titulo:
          type: string
          description: Título do chamado
        descricao:
          type: string
          description: Descrição detalhada do chamado
        dataAbertura:
          type: string
          format: date-time
          description: Data e hora de abertura do chamado no formato 'yyyy-MM-dd HH:mm:ss'
          example: "2023-12-31 23:59:59"
        status:
          type: string
          description: Status atual do chamado
        fotoAntesFile:
          type: string
          format: binary
          description: Arquivo de imagem antes da resolução (usado apenas no upload)
        fotoAntesUrl:
          type: string
          description: URL da imagem antes da resolução
        fotoDepoisUrl:
          type: string
          description: URL da imagem depois da resolução
        secretaria:
          type: string
          description: Secretaria responsável pelo chamado
        latitude:
          type: number
          format: decimal
          description: Latitude da localização do chamado
        longitude:
          type: number
          format: decimal
          description: Longitude da localização do chamado
        avaliacao:
          $ref: '#/components/schemas/AvaliacaoDTO'
        historicos:
          type: array
          items:
            $ref: '#/components/schemas/HistoricoDTO'
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'

    # Funcionario DTO
    FuncionarioDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        authId:
          type: integer
          format: int64
          description: ID do usuário de autenticação associado
        cpf:
          type: string
          description: CPF do funcionário
        cargo:
          type: string
          description: Cargo do funcionário
        secretaria:
          $ref: '#/components/schemas/Secretaria'
        dataCadastro:
          type: string
          format: date-time
          description: Data e hora de cadastro no formato 'yyyy-MM-dd HH:mm:ss'
          example: "2023-12-31 23:59:59"
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'

    # Historico DTO
    HistoricoDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        chamadoId:
          type: integer
          format: int64
          description: ID do chamado associado ao histórico
        funcionarioId:
          type: integer
          format: int64
          description: ID do funcionário que realizou a modificação
        dataModificacao:
          type: string
          format: date-time
          description: Data e hora da modificação no formato 'yyyy-MM-dd HH:mm:ss'
          example: "2023-12-31 23:59:59"
        statusAnterior:
          type: string
          description: Status anterior do chamado
        statusNovo:
          type: string
          description: Novo status do chamado
        observacao:
          type: string
          description: Observações sobre a modificação
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'

    # Usuario DTO
...
    Secretaria:
      type: string
      enum:
        - OBRAS
        - URBANISMO
      description: Secretarias disponíveis no sistema

    # Utility schemas
...
    PageInfo:
      type: object
      properties:
        size:
          type: integer
          description: Número de elementos por página
        totalElements:
          type: integer
          description: Número total de elementos
        totalPages:
          type: integer
          description: Número total de páginas
        number:
          type: integer
          description: Página atual (começando em 0)
    Link:
      type: object
      properties:
        rel:
          type: string
          description: Relação do link
        href:
          type: string
          description: URL do link